// test/fixtures/generated-sdk/users.wesley.generated.ts
/* @generated by Wesley. Do not edit. */

export interface IntentBinding {
  operation: string;
}

export type IntentKind = "NODE_ADD" | "PROPERTY_SET";

export type ObserverCardinality = "EXACTLY_ONE" | "MANY";

export type ReadingKind = "PROPERTY";

export type ValueDecoder = "STRING";

export interface MutationRegisterUserRequest {
  subject: string;
}

export type MutationRegisterUserResponse = IntentBinding;

export const mutationRegisterUserOperation = {
  operationType: "MUTATION",
  fieldName: "registerUser",
  directives: {"intent":{"kind":"NODE_ADD","subject":"subject"}},
} as const;

export type MutationRegisterUserOperation = {
  request: MutationRegisterUserRequest;
  response: MutationRegisterUserResponse;
  metadata: typeof mutationRegisterUserOperation;
};

export interface MutationAssignRoleRequest {
  subject: string;
  role: string;
}

export type MutationAssignRoleResponse = IntentBinding;

export const mutationAssignRoleOperation = {
  operationType: "MUTATION",
  fieldName: "assignRole",
  directives: {"intent":{"kind":"PROPERTY_SET","subject":"subject","key":"role","value":"role"}},
} as const;

export type MutationAssignRoleOperation = {
  request: MutationAssignRoleRequest;
  response: MutationAssignRoleResponse;
  metadata: typeof mutationAssignRoleOperation;
};

export interface QueryRoleOfRequest {
  subject: string;
}

export type QueryRoleOfResponse = string;

export const queryRoleOfOperation = {
  operationType: "QUERY",
  fieldName: "roleOf",
  directives: {"observer":{"id":"users.role-of","reading":"PROPERTY","subject":"subject","key":"role","cardinality":"EXACTLY_ONE","decoder":"STRING"}},
} as const;

export type QueryRoleOfOperation = {
  request: QueryRoleOfRequest;
  response: QueryRoleOfResponse;
  metadata: typeof queryRoleOfOperation;
};

export interface QueryRolesOfRequest {
  subjects: string[];
}

export type QueryRolesOfResponse = string[];

export const queryRolesOfOperation = {
  operationType: "QUERY",
  fieldName: "rolesOf",
  directives: {"observer":{"id":"users.roles-of","reading":"PROPERTY","subject":"subjects","key":"role","cardinality":"MANY","decoder":"STRING","maxReadings":256}},
} as const;

export type QueryRolesOfOperation = {
  request: QueryRolesOfRequest;
  response: QueryRolesOfResponse;
  metadata: typeof queryRolesOfOperation;
};